Prepare project to be resumed
- Deleted journeys code (was migrated to frontend-web, will be faster
  to migrate this back rather than edit)
- Deleted daily event code (no longer used)
- Updated to use the same directory structure as frontend-web
- Copied over new images, features router from frontend-web
- Get it running on android again

This currently implements logging in & setting a name, but then goes to
an infinite spinner. However, it has solid parity of logic flow to
the web, so I expect porting over most features to be fairly
straightforward. The intent is that all logic is kept (state, resources,
feature files) and the only difference will be in the component
implementations

Note the frontend-web currently has the login / settings screen
live outside of features since it can use standard browser routing
to get to them (window.location.href = '/settings'). I might update
it to use the normal feature strategy if it's ever a nontrivial
transformation