Silky smooth pick emotion animation
Boy was this a lot of work! This follows the performance improvements
from the web, as well as switching to a transform strategy for animating
the size changes, which means we can generally use rasterized components
rather than redrawing everything from scratch. This is critical, since the
draw time alone was taking over the render budget on my midrange
android