v83: redo login state machine management
this switches to imperative code which is easier to follow and better
handles many error cases. the specific bug being fixed that prompted
this redesign is before if resuming from a suspended state after the id
token expired it led to an infinite spinner, and before v82, it led to if
you clicked while the refresh was occurring it would switch to a logged out
state.

v82 was approved on both stores but never released due to the above bug

LoginProvider is identical (copy+paste) to the web v83 version, though may
be slightly differently formatted as the two projects have different settings